Split out api-common unit test files
Split out api-common unit tests into multiple smaller files to better readability.

Relates to review comment: https://github.com/QutEcoacoustics/workbench-client/pull/67#discussion_r368358914
